+++ nf_core/modules/modules_json.py
@@ -32,7 +32,7 @@ class ModulesJson:
         self.modules_dir = os.path.join(self.dir, "modules")
         self.modules_json = None
 
-    def create_modules_json(self):
+    def create(self):
         """
         Creates the modules.json file from the modules installed in the pipeline directory
 
@@ -225,7 +225,7 @@ class ModulesJson:
             depth += 1
         return dirs_not_covered
 
-    def modules_json_up_to_date(self):
+    def check_up_to_date(self):
         """
         Checks whether the modules installed in the directory
         are consistent with the entries in the 'modules.json' file and vice versa.
@@ -237,7 +237,7 @@ class ModulesJson:
         If a module is installed but the entry in 'modules.json' is missing we iterate through
         the commit log in the remote to try to determine the SHA.
         """
-        self.load_modules_json()
+        self.load()
         old_modules_json = copy.deepcopy(self.modules_json)
 
         # Compute the difference between the modules in the directory
@@ -263,7 +263,9 @@ class ModulesJson:
                 # If it does, check if the module is in the 'modules.json' file
                 modules_path = os.path.relpath(dir, start=module_repo_name)
 
-                if modules_path not in old_modules_json["repos"][module_repo_name].get("modules", {}):
+                if module_repo_name not in old_modules_json["repos"]:
+                    missing_from_modules_json.append(dir)
+                elif modules_path not in old_modules_json["repos"][module_repo_name].get("modules", {}):
                     missing_from_modules_json.append(dir)
                 else:
                     old_modules_json["repos"][module_repo_name]["modules"].pop(modules_path)
@@ -312,7 +314,12 @@ class ModulesJson:
                         remove_from_mod_json[repo].append(module)
                         continue
                     module_dir = [self.dir, "modules", *install_folder]
-                    modules_repo.install_module(module, module_dir, sha)
+                    if not modules_repo.install_module(module, module_dir, sha):
+                        if repo not in remove_from_mod_json:
+                            remove_from_mod_json[repo] = []
+                        log.warning(f"Could not install module '{module}' in '{repo}' - removing from modules.json")
+                        remove_from_mod_json[repo].append(module)
+                        continue
 
             # If the reinstall fails, we remove those entries in 'modules.json'
             if sum(map(len, remove_from_mod_json.values())) > 0:
@@ -455,9 +462,9 @@ class ModulesJson:
 
             shutil.move(path, local_path)
 
-        self.dump_modules_json()
+        self.dump()
 
-    def load_modules_json(self):
+    def load(self):
         """
         Loads the modules.json file into the variable 'modules_json'
 
@@ -473,7 +480,7 @@ class ModulesJson:
         except FileNotFoundError:
             raise UserWarning("File 'modules.json' is missing")
 
-    def update_modules_json(self, modules_repo, module_name, module_version, write_file=True):
+    def update(self, modules_repo, module_name, module_version, write_file=True):
         """
         Updates the 'module.json' file with new module info
 
@@ -484,7 +491,7 @@ class ModulesJson:
             write_file (bool): whether to write the updated modules.json to a file.
         """
         if self.modules_json is None:
-            self.load_modules_json()
+            self.load()
         repo_name = modules_repo.fullname
         remote_url = modules_repo.remote_url
         base_path = modules_repo.base_path
@@ -494,9 +501,9 @@ class ModulesJson:
         # Sort the 'modules.json' repo entries
         self.modules_json["repos"] = nf_core.utils.sort_dictionary(self.modules_json["repos"])
         if write_file:
-            self.dump_modules_json()
+            self.dump()
 
-    def remove_modules_json_entry(self, module_name, repo_name):
+    def remove_entry(self, module_name, repo_name):
         """
         Removes an entry from the 'modules.json' file.
 
@@ -521,7 +528,7 @@ class ModulesJson:
             log.warning(f"Module '{repo_name}/{module_name}' is missing from 'modules.json' file.")
             return False
 
-        self.dump_modules_json()
+        self.dump()
         return True
 
     def repo_present(self, repo_name):
@@ -532,6 +539,8 @@ class ModulesJson:
         Returns:
             (bool): Whether the repo exists in the modules.json
         """
+        if self.modules_json is None:
+            self.load()
         return repo_name in self.modules_json.get("repos", {})
 
     def module_present(self, module_name, repo_name):
@@ -543,6 +552,8 @@ class ModulesJson:
         Returns:
             (bool): Whether the module is present in the 'modules.json' file
         """
+        if self.modules_json is None:
+            self.load()
         return module_name in self.modules_json.get("repos", {}).get(repo_name, {}).get("modules", {})
 
     def get_modules_json(self):
@@ -553,7 +564,7 @@ class ModulesJson:
             (dict): A copy of the loaded modules.json
         """
         if self.modules_json is None:
-            self.load_modules_json()
+            self.load()
         return copy.deepcopy(self.modules_json)
 
     def get_module_version(self, module_name, repo_name):
@@ -568,7 +579,7 @@ class ModulesJson:
             (str): The git SHA of the module if it exists, None otherwise
         """
         if self.modules_json is None:
-            self.load_modules_json()
+            self.load()
         return (
             self.modules_json.get("repos", {})
             .get(repo_name, {})
@@ -588,7 +599,7 @@ class ModulesJson:
             (str): The git url of the repository if it exists, None otherwise
         """
         if self.modules_json is None:
-            self.load_modules_json()
+            self.load()
         return self.modules_json.get("repos", {}).get(repo_name, {}).get("git_url", None)
 
     def get_base_path(self, repo_name):
@@ -601,10 +612,10 @@ class ModulesJson:
             (str): The base path of the repository if it exists, None otherwise
         """
         if self.modules_json is None:
-            self.load_modules_json()
+            self.load()
         return self.modules_json.get("repos", {}).get(repo_name, {}).get("base_path", None)
 
-    def dump_modules_json(self):
+    def dump(self):
         """
         Sort the modules.json, and write it to file
         """