+++ nf_core/utils.py
@@ -117,6 +117,7 @@ class Pipeline(object):
         self.minNextflowVersion = None
         self.wf_path = wf_path
         self.pipeline_name = None
+        self.pipeline_prefix = None
         self.schema_obj = None
 
         try:
@@ -162,7 +163,7 @@ class Pipeline(object):
         """
         self.nf_config = fetch_wf_config(self.wf_path)
 
-        self.pipeline_name = self.nf_config.get("manifest.name", "").strip("'").replace("nf-core/", "")
+        self.pipeline_prefix, self.pipeline_name = self.nf_config.get("manifest.name", "").strip("'").split("/")
 
         nextflowVersionMatch = re.search(r"[0-9\.]+(-edge)?", self.nf_config.get("manifest.nextflowVersion", ""))
         if nextflowVersionMatch:
@@ -326,6 +327,7 @@ def setup_requests_cachedir():
         "backend": "sqlite",
     }
 
+    logging.getLogger("requests_cache").setLevel(logging.WARNING)
     try:
         if not os.path.exists(cachedir):
             os.makedirs(cachedir)
@@ -388,7 +390,8 @@ def poll_nfcore_web_api(api_url, post_data=None):
             else:
                 try:
                     web_response = json.loads(response.content)
-                    assert "status" in web_response
+                    if "status" not in web_response:
+                        raise AssertionError()
                 except (json.decoder.JSONDecodeError, AssertionError, TypeError) as e:
                     log.debug(f"Response content:\n{response.content}")
                     raise AssertionError(
@@ -876,9 +879,8 @@ def get_repo_releases_branches(pipeline, wfs):
 
             # Check that this repo existed
             try:
-                assert rel_r.json().get("message") != "Not Found"
-            except AssertionError:
-                raise AssertionError(f"Not able to find pipeline '{pipeline}'")
+                if rel_r.json().get("message") == "Not Found":
+                    raise AssertionError(f"Not able to find pipeline '{pipeline}'")
             except AttributeError:
                 # Success! We have a list, which doesn't work with .get() which is looking for a dict key
                 wf_releases = list(sorted(rel_r.json(), key=lambda k: k.get("published_at_timestamp", 0), reverse=True))