+++ tests/test_bump_version.py
@@ -18,7 +18,7 @@ def test_bump_pipeline_version(datafiles, tmp_path):
     # Get a workflow and configs
     test_pipeline_dir = os.path.join(tmp_path, "nf-core-testpipeline")
     create_obj = nf_core.create.PipelineCreate(
-        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir
+        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir, plain=True
     )
     create_obj.init_pipeline()
     pipeline_obj = nf_core.utils.Pipeline(test_pipeline_dir)
@@ -38,7 +38,7 @@ def test_dev_bump_pipeline_version(datafiles, tmp_path):
     # Get a workflow and configs
     test_pipeline_dir = os.path.join(tmp_path, "nf-core-testpipeline")
     create_obj = nf_core.create.PipelineCreate(
-        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir
+        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir, plain=True
     )
     create_obj.init_pipeline()
     pipeline_obj = nf_core.utils.Pipeline(test_pipeline_dir)
@@ -57,7 +57,7 @@ def test_bump_nextflow_version(datafiles, tmp_path):
     # Get a workflow and configs
     test_pipeline_dir = os.path.join(tmp_path, "nf-core-testpipeline")
     create_obj = nf_core.create.PipelineCreate(
-        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir
+        "testpipeline", "This is a test pipeline", "Test McTestFace", outdir=test_pipeline_dir, plain=True
     )
     create_obj.init_pipeline()
     pipeline_obj = nf_core.utils.Pipeline(test_pipeline_dir)