+++ tests/test_schema.py
@@ -14,6 +14,7 @@ import pytest
 import requests
 import yaml
 
+import nf_core.create
 import nf_core.schema
 
 from .utils import with_temporary_file, with_temporary_folder
@@ -26,11 +27,15 @@ class TestSchema(unittest.TestCase):
         """Create a new PipelineSchema object"""
         self.schema_obj = nf_core.schema.PipelineSchema()
         self.root_repo_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-        # Copy the template to a temp directory so that we can use that for tests
+
+        # Create a test pipeline in temp directory
         self.tmp_dir = tempfile.mkdtemp()
         self.template_dir = os.path.join(self.tmp_dir, "wf")
-        template_dir = os.path.join(self.root_repo_dir, "nf_core", "pipeline-template")
-        shutil.copytree(template_dir, self.template_dir)
+        create_obj = nf_core.create.PipelineCreate(
+            "test_pipeline", "", "", outdir=self.template_dir, no_git=True, plain=True
+        )
+        create_obj.init_pipeline()
+
         self.template_schema = os.path.join(self.template_dir, "nextflow_schema.json")
 
     def tearDown(self):